Upgrade project & dependencies (#730)
* Updates actions/cache version

Updates the `actions/cache` GitHub Action to version 4.
This ensures compatibility with the latest features and
improvements of the action.

* Upgrades dependencies and tooling

Updates project dependencies and development tools to their latest versions.

This change ensures the project benefits from the newest features,
security patches, and performance improvements in the dependency
ecosystem. It also alignes the project with the latest standards
and best practices for development workflows.

* fix safety check

* Upgrade MegaLinter

* Remove unused GITHUB_TOKEN from Dockerfile

* [MegaLinter] Apply linters fixes :)

* Use python 3.12 for tests

---------

Co-authored-by: nvuillam <17500430+nvuillam@users.noreply.github.com>